requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
openai>=1.6.1,<2.0.0
numpy==1.24.3
pandas==2.1.3
